refactor: HomeSection
