+++ src/components/HomeSection/TypingAnimation.tsx
